Reorg and refactored ThemeContext for next.js
