+++ src/contexts/theme/ThemeContext.jsx
@@ -1,4 +1,4 @@
-import React, { createContext, useState} from 'react';
+import { createContext, useContext, useState } from 'react';
 import { getThemes } from './themes';
 
 
@@ -6,11 +6,11 @@ export const ThemeContext = createContext();
 
 
 export const ThemeProvider = ({ children }) => {
-    const firstTimeThemes = getThemes().filter(theme => theme.firstTime === true), 
-          setFirstTimeTheme = firstTimeThemes[Math.floor((Math.random() * (firstTimeThemes.length)))]
+    const firstTimeThemes = getThemes().filter(theme => theme.firstTime === true),
+          setFirstTimeTheme = firstTimeThemes[Math.floor((Math.random() * (firstTimeThemes.length)))];
     const [currentTheme, setCurrentTheme] = useState(setFirstTimeTheme);
-  
-    
+
+
     const getRandomTheme = () => {
       const allThemes = getThemes();
       const themesWithoutCurrentTheme = allThemes.filter(
@@ -21,21 +21,23 @@ export const ThemeProvider = ({ children }) => {
       );
       return themesWithoutCurrentTheme[randomThemeIndex];
     }
-  
+
     const switchTheme = () => {
       const randomTheme = getRandomTheme();
       setCurrentTheme(randomTheme);
     }
 
 
-    return (
-      <ThemeContext.Provider
-        value={{
-          currentTheme,
-          switchTheme,
-        }}
-      >
-        {children}
-      </ThemeContext.Provider>
-    );
+  return (
+    <ThemeContext.Provider
+      value={{
+        currentTheme,
+        switchTheme,
+      }}
+    >{children}</ThemeContext.Provider>
+  );
+};
+
+export function useThemeContext() {
+  return useContext(ThemeContext);
 };
