+++ src/contexts/theme/index.js
@@ -1,0 +1,3 @@
+import { ThemeContext } from './ThemeContext';
+
+export { ThemeContext };
